% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic.method.R
\name{heuristic.method}
\alias{heuristic.method}
\title{Consruct a circular network use a heuristic method}
\usage{
heuristic.method(
  M,
  tree.method = "unj",
  loop.limit = 10,
  filename = "heuristic_method.nex",
  taxaname = NULL
)
}
\arguments{
\item{M}{the distance matrix for construct tree and network (the matrix should fit the triangle inequality and the diagonal should be 0).}

\item{tree.method}{method for construct the original tree for heuristic method, default is \code{unj}, for unweighted ntighbor joining tree;
\code{nj} for neighbor joining tree; \code{nnet} for symmetry nnet tree; \code{nnetns} for no symmetry nnet tree;
\code{BioNJ} for BioNJ tree.}

\item{loop.limit}{limit for the loop number of ordering improving part, default is \code{10}.
And using half the number of taxa is naturally and easily.}

\item{filename}{a character will be the naxus file's name, default is \code{heuristic_method.nex}.}

\item{taxaname}{a character set of names for taxa, ordering is consist with original distance matrix \code{M}.}
}
\value{
The LSfit value.
}
\description{
For the process, the heuristic method function is similar with \code{\link{lpnet}}.
The only different is the step which calculate the circular ordering from a tree.
The heuristic method flip the edge from top to bottom which has the biggest \code{W} value.
And the \code{W} value is the average of all changed quartets weights over the changed quartets number if flip the edge.
Finally, we cycle through an improvement step that checks all edges and flips which edges can improve the sum of all quartets
until no improvement or the number of cycles reaches the loop limit.
Actually, \code{\link{lpnet}} has higher accuracy than heuristic method.
As an alternative to \code{\link{lpnet}}, the heuristic method can handle more taxa number.
}
\examples{
### From Huson and Bryant (2006, Fig 4):
x <- c(14.06, 17.24, 20.5, 23.37, 17.43, 19.18, 18.48, 9.8, 13.06, 15.93, 15.65,
       17.4, 16.7, 6.74, 16.87, 16.59, 18.34, 17.64, 17.57, 17.29, 19.04, 18.34,
       17.6, 19.35, 21.21, 9.51, 11.37, 13.12)
M <- matrix(0, 8, 8)
M[row(M) > col(M)] <- x
M <- M + t(M)
taxaname <- c("A", "B", "C", "D", "E", "F", "G", "H")
heuristic.method(M,
                 tree.method = "nj",
                 loop.limit = 4,
                 filename = "example.nex",
                 taxaname = taxaname)

}
