\name{lpNet-package}
\alias{lpNet-package}
\alias{lpNet}
\docType{package}
\title{
Network Inference Of Perturbation Data Using a Linear Programming Approach.
}
\description{
lpNet takes perturbation data as input and generates an LP model which allows the inference of signaling networks. For parameter identification either leave-one-out cross-validation or stratified n-fold cross-validation can be used.
}
\details{
\tabular{ll}{
Package: \tab lpNet\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-11\cr
License: \tab Artistic License 2.0\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
% B. Knapp, J. Mazur, L. Kaderali
% Maintainer: bettina.knapp@tu-dresden.de
}
\references{
B. Knapp, RNA Interference Data: from a Statistical Analysis to Network Inference, PhD Thesis, 2012, \url{http://archiv.ub.uni-heidelberg.de/volltextserver/13322/}
%J. Mazur, D. Ritter, G. Reinelt, L. Kaderali, Reconstructing nonlinear dynamic models of gene regulation using stochastic sampling. BMC Bioinformatics, 10:448, 2009.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ network inference }
\keyword{linear programming approach}
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
