% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpnet.input.tree.R
\name{lpnet.input.tree}
\alias{lpnet.input.tree}
\title{Use an Input Tree for Lpnet}
\usage{
lpnet.input.tree(
  M,
  tree,
  lp.package = "Rglpk",
  lp.type = NULL,
  filename = "lpnet.nex",
  taxaname = NULL
)
}
\arguments{
\item{M}{the distance matrix for construct network (the matrix should fit the triangle inequality and the diagonal should be 0).}

\item{tree}{the phylogenetic tree which class is \code{phylo} for lpnet algorithm (the \code{edge} block is necessary).}

\item{lp.package}{which package will used for Linear Programming, default is \code{Rglpk}, for a free R package;
\code{gurobi} for the gurobi package.}

\item{lp.type}{a character vector indicating the types of the objective variables. default is \code{NULL}, for ordinary;
\code{B} for binary; \code{I} for intrger; \code{C} for continuous.}

\item{filename}{a character will be the naxus file's name, default is \code{lpnet.nex}.}

\item{taxaname}{a character set of names for taxa, ordering is consist with original distance matrix \code{M}.}
}
\value{
The LSfit value.
}
\description{
Replacing the tree which constructed by the methods from \code{\link{lpnet}} to an arbitrary input tree which need to be input.
}
\examples{
### From Huson and Bryant (2006, Fig 4):
x <- c(14.06, 17.24, 20.5, 23.37, 17.43, 19.18, 18.48, 9.8, 13.06, 15.93, 15.65,
       17.4, 16.7, 6.74, 16.87, 16.59, 18.34, 17.64, 17.57, 17.29, 19.04, 18.34,
       17.6, 19.35, 21.21, 9.51, 11.37, 13.12)
M <- matrix(0, 8, 8)
M[row(M) > col(M)] <- x
M <- M + t(M)
njtree <- ape::nj(M)
taxaname <- c("A", "B", "C", "D", "E", "F", "G", "H")
lpnet.input.tree(M,
                 tree = njtree,
                 lp.package = "Rglpk",
                 lp.type = NULL,
                 filename = "example.nex",
                 taxaname = taxaname)

}
