% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpnet.R
\name{lpnet}
\alias{lpnet}
\title{Consruct a circular network based on a tree by Linear Programming}
\usage{
lpnet(
  M,
  tree.method = "unj",
  lp.package = "Rglpk",
  lp.type = NULL,
  filename = "lpnet.nex",
  taxaname = NULL,
  sequencelength = 1
)
}
\arguments{
\item{M}{the distance matrix for construct tree and network;
(the matrix should fit the triangle inequality and the diagonal should be 0).}

\item{tree.method}{method for construct the original tree for lp, default is \code{unj}, for unweighted ntighbor joining tree;
\code{nj} for neighbor joining tree; \code{nnet} for symmetry nnet tree; \code{nnetns} for no symmetry nnet tree;
\code{BioNJ} for BioNJ tree.}

\item{lp.package}{which package will used for Linear Programming, default is \code{Rglpk}, for a free R package;
\code{gurobi} for the gurobi package.}

\item{lp.type}{a character vector indicating the types of the objective variables. default is \code{NULL}, for ordinary;
\code{B} for binary; \code{I} for intrger; \code{C} for continuous.}

\item{filename}{a character will be the naxus file's name, default is \code{lpnet.net}}

\item{taxaname}{a character set of names for taxa, ordering is consist with original distance matrix \code{M}.}

\item{sequencelength}{the sequence length of the data only used for BioNJ (default is 1 for only distance matrix).}
}
\value{
None (invisible ‘NULL’).
}
\description{
Construct a planner network which has a circular ordering for a distance matrix, and write a nexus file for SplitsTree4.
First construct a tree for the distance matrix.Then use Linear Programming(lp) to change the circular ordering.
The ordering have the biggest sum of quartets for all taxa is the lp net ordering.
Then use Non-negative least squares(nnls) to calculate weights of splits which are consist with the lp net ordering.
Finally, write a nexus file with taxa block and splits block for SplitsTree4 to see the circular network.
}
\examples{
### From Huson and Bryant (2006, Fig 4):
x <- c(14.06, 17.24, 20.05, 23.37, 17.43, 19.18, 18.48, 9.8, 13.06, 15.93, 15.65,
       17.4, 16.7, 6.74, 16.87, 16.59, 18.34, 17.64, 17.57, 17.29, 19.04, 18.34,
       17.6, 19.35, 21.21, 9.51, 11.37, 13.12)
M <- matrix(0, 8, 8)
M[row(M) > col(M)] <- x
M <- M + t(M)
taxaname<-c("A", "B", "C", "D", "E", "F", "G", "H")
lpnet(M,
      tree.method = "nj",
      lp.package = "Rglpk",
      lp.type = NULL,
      filename = "example.nex",
      taxaname = taxaname)

}
