% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nexus.R
\name{read.nexus.distanceblock}
\alias{read.nexus.distanceblock}
\title{Read the Distance block from a Nexus file}
\usage{
read.nexus.distanceblock(filename, n)
}
\arguments{
\item{filename}{a character, the name of the file which will be read for the distance matrix.}

\item{n}{the number of taxa}
}
\value{
A symmetry distance matrix
}
\description{
Read the Distance block from a nexus file, and get the symmetry distance matrix.
(The distance block should be set as \code{diagonal} and \code{triangle=both}).
}
\seealso{
\code{\link{read.nexus.taxablock}}
}
