% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nexus.R
\name{read.nexus.taxablock}
\alias{read.nexus.taxablock}
\title{Read the Taxa block from a Nexus file}
\usage{
read.nexus.taxablock(filename, n)
}
\arguments{
\item{filename}{a character, the name of the file which will be read for the taxa labels.}

\item{n}{the number of taxa}
}
\value{
A character vector of the taxa's labels read from the nexus file.
}
\description{
Read the taxa block from a nexus file, and get the taxa's labels.
}
\seealso{
\code{\link{read.nexus.distanceblock}}
}
