\name{lpNet-package}
\alias{lpNet-package}
\alias{lpNet}
\docType{package}
\title{
Network Inference Of Perturbation Data Using a Linear Programming Approach.
}
\description{
lpNet aims at infering biological networks, in particular signaling and gene networks.
For that it takes perturbation data, either steady-state or time-series, 
as input and generates an LP model which allows the inference of signaling networks. 
For parameter identification either leave-one-out cross-validation or 
stratified n-fold cross-validation can be used.
}
\details{
\tabular{ll}{
Package: \tab lpNet\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-11\cr
License: \tab Artistic License 2.0\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
B. Knapp, M. R. A. Matos, J. Mazur, L. Kaderali

Maintainer: bettina.knapp@helmholtz-muenchen.de
}
\references{
Bettina Knapp and Lars Kaderali, Reconstruction of cellular signal transduction networks using perturbation assays and linear programming, PLoS ONE, 2013.

Marta R. A. Matos, Network inference : extension of linear programming model for time-series data, Master's thesis, Department of Informatics, University of Minho.
%J. Mazur, D. Ritter, G. Reinelt, L. Kaderali, Reconstructing nonlinear dynamic models of gene regulation using stochastic sampling. BMC Bioinformatics, 10:448, 2009.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ network inference }
\keyword{linear programming approach}
